[IMP] l10n_es{,edi_sii,edi_tbai}: move some code to l10n_es
The moved code will also be needed for Veri*Factu

- patched http adapter
- a function to retrieve partner info

part of
task-3745982
