+++ addons/l10n_es/models/__init__.py
@@ -1,6 +1,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import account_move
 from . import account_tax
+from . import res_partner
 from . import template_es_assec
 from . import template_es_common
 from . import template_es_coop_full
