+++ addons/l10n_es/tools/__init__.py
@@ -1,0 +1,1 @@
+from . import http_adapter
