+++ addons/l10n_es/tools/http_adapter.py
@@ -1,0 +1,47 @@
+import requests
+
+from urllib3.util.ssl_ import create_urllib3_context
+from urllib3.contrib.pyopenssl import inject_into_urllib3
+from OpenSSL.crypto import load_certificate, load_privatekey, FILETYPE_PEM
+
+# Custom patches to perform the WSDL requests.
+# Avoid failure on servers where the DH key is too small
+EUSKADI_CIPHERS = "DEFAULT:!DH"
+
+
+class PatchedHTTPAdapter(requests.adapters.HTTPAdapter):
+    """ An adapter to block DH ciphers which may not work for the tax agencies called"""
+
+    def init_poolmanager(self, *args, **kwargs):
+        # OVERRIDE
+        inject_into_urllib3()
+        kwargs['ssl_context'] = create_urllib3_context(ciphers=EUSKADI_CIPHERS)
+        return super().init_poolmanager(*args, **kwargs)
+
+    def cert_verify(self, conn, url, verify, cert):
+        # OVERRIDE
+        # The last parameter is only used by the super method to check if the file exists.
+        # In our case, cert is an odoo record 'l10n_es_edi.certificate' so not a path to a file.
+        # By putting 'None' as last parameter, we ensure the check about TLS configuration is
+        # still made without checking temporary files exist.
+        super().cert_verify(conn, url, verify, None)
+        conn.cert_file = cert
+        conn.key_file = None
+
+    def get_connection(self, url, proxies=None):
+        # OVERRIDE
+        # Patch the OpenSSLContext to decode the certificate in-memory.
+        conn = super().get_connection(url, proxies=proxies)
+        context = conn.conn_kw['ssl_context']
+
+        def patched_load_cert_chain(l10n_es_odoo_certificate, keyfile=None, password=None):
+            cert_file, key_file, _certificate = l10n_es_odoo_certificate.sudo()._decode_certificate()
+            cert_obj = load_certificate(FILETYPE_PEM, cert_file)
+            pkey_obj = load_privatekey(FILETYPE_PEM, key_file)
+
+            context._ctx.use_certificate(cert_obj)
+            context._ctx.use_privatekey(pkey_obj)
+
+        context.load_cert_chain = patched_load_cert_chain
+
+        return conn
