+++ addons/l10n_es_edi_tbai/models/account_edi_format.py
@@ -17,7 +17,7 @@ from pytz import timezone
 from requests.exceptions import RequestException
 
 from odoo import _, models, release
-from odoo.addons.l10n_es_edi_sii.models.account_edi_format import PatchedHTTPAdapter
+from odoo.addons.l10n_es.tools.http_adapter import PatchedHTTPAdapter
 from odoo.addons.l10n_es_edi_tbai.models.l10n_es_edi_tbai_agencies import get_key
 from odoo.addons.l10n_es_edi_tbai.models.xml_utils import (
     NS_MAP, bytes_as_block, calculate_references_digests,
